Move createGuid function to separate file
